backend/observability/requirements.txt
# Python dependencies for The HIVE Observability Infrastructure

# Core frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client and async support
aiohttp==3.9.1
httpx==0.25.2

# Observability and monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-aiohttp-client==0.42b0

# Audio processing for synthetic load testing
numpy==1.24.4
soundfile==0.12.1
scipy==1.11.4

# System monitoring
psutil==5.9.6
GPUtil==1.4.0

# Redis client
redis==5.0.1

# WebSocket support for real-time metrics
websockets==12.0

# Data processing
pandas==2.1.4
matplotlib==3.8.2

# Logging and utilities
structlog==23.2.0
python-json-logger==2.0.7